Move all videos swr fetching to PringPringCatsIndex
